Move project select to header
